website/requirements.txt
Flask
numpy
Pillow
